New TS with TS errors resolved
